pyproject.toml
[project]
name = "tickethub-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "channels[daphne]>=4.2.2",
    "django>=5.2",
    "django-markdownfield>=0.11.0",
    "django-silk>=5.3.2",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "markdown>=3.8",
    "pillow>=11.2.1",
    "python-decouple>=3.8",
]
